가격순필터링,카테고리선택비동기화

+++ src/main/java/com/smhrd/mueossa/dto/ProductAndCategoryDTO.java
@@ -20,7 +20,7 @@ public class ProductAndCategoryDTO {
   private String pImage;
   private String pGroup;
   private String pType;
-  private int sentiment;
+  private Float sentiment;
 
   private String ctComf;
